api/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
certifi==2023.5.7; python_version >= '3.6'
charset-normalizer==2.0.12; python_version >= '3.5'
click==8.1.3; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
dnspython==2.3.0; python_version >= '3.7' and python_version < '4.0'
exchange-calendars==4.2.8; python_version ~= '3.8'
flask-cors==3.0.10
flask==2.1.2
gunicorn==20.1.0
idna==3.4; python_version >= '3.5'
importlib-metadata==6.6.0; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
joblib==1.2.0; python_version >= '3.7'
korean-lunar-calendar==0.3.1
markupsafe==2.1.3; python_version >= '3.7'
numpy==1.24.3; python_version >= '3.8'
packaging==23.1; python_version >= '3.7'
pandas-market-calendars==3.5
pandas==2.0.2
patsy==0.5.3
pyluach==2.2.0; python_version >= '3.7'
pymongo[srv]==4.3.3
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.20.0
pytz==2023.3
requests==2.28.0
scikit-learn==1.2.2
scipy==1.10.1; python_version < '3.12' and python_version >= '3.8'
setuptools==67.8.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
statsmodels==0.14.0
threadpoolctl==3.1.0; python_version >= '3.6'
toolz==0.12.0; python_version >= '3.5'
tzdata==2023.3; python_version >= '2'
urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.3.4; python_version >= '3.8'
zipp==3.15.0; python_version >= '3.7'
